requirements.txt
bleach==2.0.0
Django==1.11.4
django-markdownify==0.2.1
django-model-utils==3.0.0
django-recaptcha==1.3.1
easy-thumbnails==2.4.1
html5lib==0.999999999
Markdown==2.6.9
olefile==0.44
Pillow==4.2.1
pytz==2017.2
six==1.10.0
sorl-thumbnail==12.3
webencodings==0.5.1
